# admin-api/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Unchain Admin API
  description: |
    Comprehensive OpenAPI specification for the Unchain Admin API.

    This API controls the configuration of the Unchain Feature Management platform, 
    including Features, Strategies, Projects, Environments, Context Fields, 
    Change Requests, Segments, FeatureTags, and Dependencies.

    **Base URL**: `<your-unchain-instance>/api/admin`
  version: 4.0.0
  contact:
    name: Philipp Walther
servers:
  - url: http://localhost:8080
    description: Local Instance
security:
  - OidcAuth: []
...
paths:
...
  /projects/{projectId}:
    get:
      operationId: getProject
      x-required-permissions: [PROJECT_READ]
      summary: Get project details
      tags: [Projects]
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    put:
      operationId: updateProject
      x-required-permissions: [PROJECT_UPDATE]
      summary: Update project
      tags: [Projects]
      parameters:
        - $ref: "#/components/parameters/projectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Project updated
    delete:
      operationId: deleteProject
      x-required-permissions: [PROJECT_DELETE]
      summary: Delete project
      tags: [Projects]
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: Project deleted
  /projects/{projectId}/features:
    get:
      operationId: getFeaturesByProject
      x-required-permissions: [PROJECT_READ]
      summary: Get all features for a project
      tags: [Features]
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: List of features
          content:
            application/json:
              schema:
                type: object
                properties:
                  features:
                    type: array
                    items:
                      $ref: "#/components/schemas/Feature"
    post:
      operationId: createFeature
      x-required-permissions: [FEATURE_CREATE]
      summary: Create a new feature toggle
      tags: [Features]
      parameters:
        - $ref: "#/components/parameters/projectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeatureRequest"
      responses:
        "201":
          description: Feature created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feature"
  /projects/{projectId}/features/{featureName}:
    get:
      operationId: getFeature
      x-required-permissions: [FEATURE_READ]
      summary: Get a specific feature
      tags: [Features]
      parameters:
        - $ref: "#/components/parameters/projectId"
        - $ref: "#/components/parameters/featureName"
      responses:
        "200":
          description: Feature details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feature"
    put:
      operationId: updateFeature
      summary: Update a feature toggle
      tags: [Features]
      parameters:
        - $ref: "#/components/parameters/projectId"
        - $ref: "#/components/parameters/featureName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeatureRequest"
      responses:
        "200":
          description: Feature updated
    delete:
      operationId: deleteFeature
      x-required-permissions: [FEATURE_DELETE]
      summary: Archive a feature toggle
      tags: [Features]
      parameters:
        - $ref: "#/components/parameters/projectId"
        - $ref: "#/components/parameters/featureName"
      responses:
        "200":
          description: Feature archived successfully

  # =========================================================================
  # FEATURE STRATEGIES & ENVIRONMENTS
  # =========================================================================
...
  /projects/{projectId}/features/{featureName}/environments/{environment}/on:
    post:
      operationId: enableFeature
      x-required-permissions: [FEATURE_UPDATE]
      summary: Enable feature in environment
      tags: [Features]
      parameters:
        - $ref: "#/components/parameters/projectId"
        - $ref: "#/components/parameters/featureName"
        - $ref: "#/components/parameters/environment"
      responses:
        "200":
          description: Feature enabled
  /projects/{projectId}/features/{featureName}/environments/{environment}/off:
    post:
      operationId: disableFeature
      x-required-permissions: [FEATURE_UPDATE]
      summary: Disable feature in environment
      tags: [Features]
      parameters:
        - $ref: "#/components/parameters/projectId"
        - $ref: "#/components/parameters/featureName"
        - $ref: "#/components/parameters/environment"
      responses:
        "200":
          description: Feature disabled

  # =========================================================================
  # DEPENDENCIES
  # =========================================================================
...
components:
  securitySchemes:
    OidcAuth:
      type: openIdConnect
      openIdConnectUrl: https://server.redmoon.ch/auth/oidc/.well-known/openid-configuration
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
      description: The ID of the project (e.g. 'default')
    featureName:
      name: featureName
      in: path
      required: true
      schema:
        type: string
      description: The name of the feature toggle
    environment:
      name: environment
      in: path
      required: true
      schema:
        type: string
      description: The environment name (e.g. 'development', 'production')
...
  schemas:
...
    Feature:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
        project:
          type: string
        stale:
          type: boolean
        createdAt:
          type: string
          format: date-time
        impressionData:
          type: boolean
          description: Whether to collect impression data for this feature
        environments:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/FeatureEnvironment"
        tags:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/FeatureTag"
    CreateFeatureRequest:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [release, experiment, operational, kill-switch, permission]
        impressionData:
          type: boolean
          description: Whether to collect impression data for this feature
    UpdateFeatureRequest:
      type: object
      properties:
        description:
          type: string
        type:
          type: string
        stale:
          type: boolean
        impressionData:
          type: boolean
    FeatureEnvironment:
      type: object
      properties:
        name:
          type: string
        enabled:
          type: boolean
        strategies:
          type: array
          items:
            $ref: "#/components/schemas/Strategy"

    # --- Strategy Schemas ---
    Strategy:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        constraints:
          type: array
          items:
            $ref: "#/components/schemas/Constraint"
        editable:
          type: boolean
          description: Whether this strategy can be edited (predefined strategies cannot be edited)
        deprecated:
          type: boolean
          description: "Whether this strategy is deprecated (note: predefined strategies cannot be deprecated)"
        parameters:
          type: array
          items:
            $ref: "#/components/schemas/StrategyParameter"
        segments:
          type: array
          items:
            type: integer
            description: IDs of segments applied to this strategy
    StrategyParameter:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
          description: The name of the parameter.
          example: percentage
        type:
          type: string
          description: The type of the parameter.
          enum:
            - string
            - percentage
            - list
            - number
            - boolean
          example: percentage
        description:
          type: string
          description: A description of this strategy parameter.
          example: How many percent of users should see this feature?
        required:
          type: boolean
          description: Whether this parameter must be configured when using the strategy.
          default: false
          example: false
        value:
          type: string
          description: The value of the parameter when used in a strategy instance.
...
    Constraint:
      type: object
      properties:
        contextName:
          type: string
        operator:
          type: string
          enum:
            [
              IN,
              NOT_IN,
              STR_ENDS_WITH,
              STR_STARTS_WITH,
              STR_CONTAINS,
              NUM_EQ,
              NUM_GT,
              NUM_LT,
              DATE_AFTER,
              DATE_BEFORE,
              SEMVER_EQ,
              SEMVER_GT,
              SEMVER_LT,
            ]
        values:
          type: array
          items:
            type: string
        caseInsensitive:
          type: boolean
        inverted:
          type: boolean

    # --- Project Schemas ---
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        health:
          type: integer
        updatedAt:
          type: string
          format: date-time
...
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
...
    FeatureTag:
      type: object
      required:
        - type
        - value
      properties:
        type:
          type: string
          example: "simple"
        value:
          type: string
          example: "payment-team"
...